Add handling for None values in SourceGroups
This allows some outputs in a MatrixGroup not to have a
target for certain signal sets, his is useful in for instance
a multi-head KVM in there are signal head source PCs. None
can be added to the sources list in the group to indicate
any 'holes' in the list but they are not consider 'companion'
sources.